Rename package to pkg
